fixed again.
git-svn-id: svn://svn.coreboot.org/coreboot/trunk@1619 2b7e53f0-3cfb-0310-b3e9-8179ed1497e1

+++ src/northbridge/via/vt8601/raminit.c
@@ -88,7 +88,7 @@ void setnorthb(device_t north, uint8_t reg, uint8_t val)
 void
 dumpnorth(device_t north) 
 {
-	uint8_t r, c;
+	unsigned int r, c;
 	for(r = 0; ; r += 16) {
 		print_debug_hex8(r);
 		print_debug(":");
